+++ R/reexports.R
@@ -1,22 +1,3 @@
-
-# ----------------------- groupedstats -----------------------------------
-
-#' @export
-#' @importFrom groupedstats specify_decimal_p
-groupedstats::specify_decimal_p
-
-#' @export
-#' @importFrom groupedstats signif_column
-groupedstats::signif_column
-
-#' @export
-#' @importFrom groupedstats lm_effsize_ci
-groupedstats::lm_effsize_ci
-
-#' @export
-#' @importFrom groupedstats set_cwd
-groupedstats::set_cwd
-
 # ----------------------- tidyverse -------------------------------------
 
 #' @export